Airbase-ng: Cast value to float (CID 1309287).

+++ src/airbase-ng.c
@@ -3945,9 +3945,9 @@ void cfrag_thread( void )
 		* (float) ( tv2.tv_sec  - tv.tv_sec  )
                     + (float) ( tv2.tv_usec - tv.tv_usec );
 
-        ticks[0] += f / ( 1000000/RTC_RESOLUTION );
-        ticks[1] += f / ( 1000000/RTC_RESOLUTION );
-        ticks[2] += f / ( 1000000/RTC_RESOLUTION );
+        ticks[0] += f / ((float)( 1000000/RTC_RESOLUTION ));
+        ticks[1] += f / ((float)( 1000000/RTC_RESOLUTION ));
+        ticks[2] += f / ((float)( 1000000/RTC_RESOLUTION ));
 
         if( ( (double)ticks[2] / (double)RTC_RESOLUTION )  >= ((double)
 #if defined(__x86_64__) && defined(__CYGWIN__)
